Remove libkpod. Replace runtime generation function.
