+++ cmd/podman/common.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"os"
 	"reflect"
 	"regexp"
 	"strings"
@@ -9,7 +8,6 @@ import (
 	"github.com/containers/storage"
 	"github.com/fatih/camelcase"
 	"github.com/pkg/errors"
-	"github.com/projectatomic/libpod/libkpod"
 	"github.com/projectatomic/libpod/libpod"
 	"github.com/urfave/cli"
 )
@@ -23,68 +21,8 @@ var (
 )
 
 const (
-	crioConfigPath = "/etc/crio/crio.conf"
-	idTruncLength  = 12
+	idTruncLength = 12
 )
-
-func getRuntime(c *cli.Context) (*libpod.Runtime, error) {
-
-	config, err := getConfig(c)
-	if err != nil {
-		return nil, errors.Wrapf(err, "could not get config")
-	}
-
-	options := storage.DefaultStoreOptions
-	options.GraphRoot = config.Root
-	options.RunRoot = config.RunRoot
-	options.GraphDriverName = config.Storage
-	options.GraphDriverOptions = config.StorageOptions
-
-	return libpod.NewRuntime(libpod.WithStorageConfig(options), libpod.WithConmonPath(config.Conmon), libpod.WithOCIRuntime(config.Runtime), libpod.WithCNIConfigDir(config.NetworkDir))
-}
-
-func getConfig(c *cli.Context) (*libkpod.Config, error) {
-	config := libkpod.DefaultConfig()
-	var configFile string
-	if c.GlobalIsSet("config") {
-		configFile = c.GlobalString("config")
-	} else if _, err := os.Stat(crioConfigPath); err == nil {
-		configFile = crioConfigPath
-	}
-	// load and merge the configfile from the commandline or use
-	// the default crio config file
-	if configFile != "" {
-		err := config.UpdateFromFile(configFile)
-		if err != nil {
-			return config, err
-		}
-	}
-	if c.GlobalIsSet("root") {
-		config.Root = c.GlobalString("root")
-	}
-	if c.GlobalIsSet("runroot") {
-		config.RunRoot = c.GlobalString("runroot")
-	}
-	if c.GlobalIsSet("conmon") {
-		config.Conmon = c.GlobalString("conmon")
-	}
-	if c.GlobalIsSet("storage-driver") {
-		config.Storage = c.GlobalString("storage-driver")
-	}
-	if c.GlobalIsSet("storage-opt") {
-		opts := c.GlobalStringSlice("storage-opt")
-		if len(opts) > 0 {
-			config.StorageOptions = opts
-		}
-	}
-	if c.GlobalIsSet("runtime") {
-		config.Runtime = c.GlobalString("runtime")
-	}
-	if c.GlobalIsSet("cni-config-dir") {
-		config.NetworkDir = c.GlobalString("cni-config-dir")
-	}
-	return config, nil
-}
 
 func splitCamelCase(src string) string {
 	entries := camelcase.Split(src)
